Refactor & final fixes
