+++ client/src/components/Layout/SidebarBig.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 
 import styles from './Sidebar.module.less';
 import SidebarItem from './SidebarItem';
-import Logo from '../../../public/wosp.svg';
+import Logo from '@/assets/wosp.svg';
 
 const { Header, Footer } = Layout;
