+++ client/src/components/forms/BoxToSettleForm/BoxToSettleForm.tsx
@@ -2,8 +2,13 @@ import React from 'react';
 import { FormButton, FormHOC } from '@/components';
 
 export const BoxToSettleForm = () => {
+  const onSubmit = () => {
+    // TODO: Navigate to another page?
+    return;
+  };
+
   return (
-    <FormHOC name="boxToSettleForm">
+    <FormHOC onFinish={onSubmit} name="boxToSettleForm">
       <FormButton htmlType="submit" type="primary">
         Jestem gotowy rozliczyć następną puszkę
       </FormButton>
